Fixup tests for >= 5.8
In 5.8 the place where the values are all created is reported at
line 3 rather than spread over line 2 and 3.

On revisiting these tests they're structured quite oddly in order to be
confident that the output is just what is expected and no more, they
should ideally be simplified in the future.